Resolve circular include by extracting client config
