+++ include/ua_server.h
@@ -20,13 +20,14 @@ extern "C" {
 #include "ua_types.h"
 #include "ua_types_generated.h"
 #include "ua_types_generated_handling.h"
-#include "ua_client.h"
 
 struct UA_ServerConfig;
 typedef struct UA_ServerConfig UA_ServerConfig;
 
 struct UA_Server;
 typedef struct UA_Server UA_Server;
+
+struct UA_ClientConfig;
 
 /**
  * .. _server:
@@ -540,13 +541,6 @@ UA_Server_setServerOnNetworkCallback(UA_Server *server,
                                      void* data);
 
 #endif /* UA_ENABLE_DISCOVERY_MULTICAST */
-
-/* Get the client configuration from the configuration plugin. Used by the
- * server when it needs client functionality to register to a discovery server.
- *
- * @return The client configuration structure */
-UA_ClientConfig UA_EXPORT
-UA_Server_getClientConfig(void);
 
 #endif /* UA_ENABLE_DISCOVERY */
